daemon: port to C using ctlra midi io
This commit removes any C++ dependencies from the daemon
example, instead using the Ctlra built-in MIDI I/O mechanics,
which are currently implemented for ALSA only. The basic MIDI
IO offered is adequate for the use, and can easily be ported
to other OSs using #defines.
